Accept values on instantiation only
There are multiple methods which need to know the values in order to replace
placeholders or list missing ones, and passing the values multiple times if
possible.

Therefore it’s useful to pass the values only once. But modifying the state of
an object is best avoided if possible. So the best time to pass the values is
on instantiation.