Rename from `Placeholders` to `Template`
In trying to use this class I found it clumsy, because it accepted one attribute
of a template object (`content`) but not the others (`name`, `id`).

This commit changes the class to accept an entire template object (with `name`
and `id` still optional). All attributes of the template are passed through, and
the additional properties available (eg `template.replaced`) stay the same,
although in some cases are renamed for clarity.

This commit also updates the documentation to match.