Add option to drop values before parsing template
A user’s uploaded data may well have fields which aren’t in the template,
for example `phone number` wouldn’t be in the template, but would need to be
in the CSV. You wouldn’t want to raise an error when giving the template this
data.

This commit adds an option, when instantiating the `Template` object, to drop
values from the user’s data before attempting to do any replacing of the
template.