Optionally prefix templates with service name
All text messages should be prefixed with the name of the service, eg:
> Vehicle tax: Due now. Renew online at…

This commit adds an optional `prefix` parameter to the `Template` class through
which the service name can be passed. When the placeholders are formatted or
replaced the prefix will automatically be added.

For example:
```python
  template = Template({"content": "Hello world"}, prefix="My service")
  template.formatted
  >>> Markup(u"My service: Hello world")
```

`template.content` still returns the template _without_ the prefix, for when
you need to repopulate a form, for example.