Rewrite function as a generator function
Generators are usually better because they save having to keep the whole list
(in the case the recipients) in memory.

The Python `csv` library can behave as an iterator and return its result
line-by-line as long as the data that it’s being passed does the same
(https://docs.python.org/2/library/csv.html#csv.reader).

Since we’re handing off to Celery on an recipient-by-recipient basis we can
avoid ever having to keep the whole list of recipients in memory, which _should_
make it more efficient to handle large lists.