Handle CSVs with email addresses or phone numbers
This commit adds a lookup for determining which column of a CSV contains the
recipient of a message.

The reason for doing this is that ‘mobile number’ is a much better column name
than ‘to’ if you’re a user who’s struggling to understand what’s going on with
these CSV things.