Refactor existing CSV code into a class
Since we’re going to add a lot more functions here, which will all need the same
data, the CSV processing module should be a class. The main data that the
methods will need to share is the contents of the CSV file.

The existing methods for getting recipients and personalisation can now be
properties, rather than plain functions.

This commit also renames the module for process_csv to recipients, so we can
add code that is not directly related to CSV processing.