Adds methods for getting annotated rows
Rather than try to work out which rows have errors in the admin app templates,
this commit adds a method which:
- takes the rows of a CSV
- reformats each field as a `dict`, with properties that will be useful for
  displaying the contents of the CSV back to the user

Also, the admin app probably wont want to show the entire CSV when it gets very
long, so there is another method for getting the above, but with limits on how
many rows are returned.

This should also keep performance fairly decent because all the methods are
written as generators, so only the minimum rows needed are ever read into
memory.