Adds methods for getting rows with bad data
There are multiple different ways in which a CSV can be ‘bad’. We want to
provide as helpful error messages as possible, so we need methods for all the
possible errors. This commit also adds more generic error methods, which work
by returning a combination of the more specific errors.