Don’t mix errors with normal rows
Showing a preview of the CSV file where errors are intermingled with normal
rows isn’t that clear. The reckon was that it would be good for context, but
it’s probably clearer to only show the rows with errors.

This means that the util for processing CSVs can be made simpler, because its
methods only have to return rows with errors, or all rows, never a mixture of
the two. Which is why this commit is able to delete a load of tests.

Major version bump because this removes/renames a bunch of methods.