Strip trailing empty rows from CSV files
Numbers on OS X will include in its export any empty cells which are
displayed on screen. This results in a CSV file with lots of this at the
end of it:

```
,,,
,,,
,,,
```

Currently the CSV validation throws an error, saying that data is
missing for these rows.

Instead it should just ignore them.

This commit add test data and a fix in order to do so.