Change data structure of rows to contain columns
Previously a row was a dictionary, keyed by the columns. It also had an
extra key, `index`, which said which row number it was in the file.

For example:
```
{
    'name': 'Chris',
    'day': 'Friday',
    'index': 0
}
```

If your CSV file looked like this, however:
```
phone number, index
07815838437, 27364
```
…you would lose the data in the index column.

This will only get worse if we add additional properties to the row
which might clash with the name of the columns. So this commit changes
the structure of the row dict to contain the columns, ie:
```
{
    'columns': {
        'name': 'Chris',
        'day': 'Friday',
    },
    'index': 0
}
```

This commit represents a breaking change to any users of
`RecipientCSV.get_annotated_rows`