Add concept of limit to the template class
The template class contains all the information it needs to know if a
message is too long, except for what the limit is.

This commit allows the template class to be passed a
`content_character_limit` parameter, and adds a `.content_too_long`
property which returns `True` or `False` depending on whether the final
message (with prefix and placeholders replaced) is longer than the
limit.

This saves having to repeat this logic multiple times in the apps.