Annotate rows that have overly-long content
Rows returned by the `RecipientCSV` class are annotated, currently just
with their index.

This commit adds an extra annotation which specifices whether processing
that row would result in a overly-long message. What counts as being
overly-long is determined by the instance of `Template` that is passed
to the constructor.

If no instance is passed, the rows default to *not* resulting in
overly-long messages.