Ignore case an spacing in CSV file column headers
Users often won’t have control over the data that they are uploading.
But we mandate, for example that their recipient column is `email
address` or `phone number` with that exact case. This is also a problem
if they put a trailing space in one of their placeholders, or use Title
Case when their file has UPPERCASE columns, etc.

_However_, we don’t want to change how we show column headers on the
front end (should match what the user uploaded).

So this commit replaces the `dict` of columns that we return for each
row with an object that:
- behaves like a `dict`, ie you can do `.get()` on it and, subscript it
- ignores case and spacing, so ‘phone number’, ‘PHONENUMBER’ and ‘
  Phone_number  ` will all give the same value