Add an `allowed_to_send_to` property to a CSV
It is useful to whether, as a whole, a user is allowed to send to
all the recipients in a CSV file. It will help us give better error
messages on the front end, without having to look through each row
in the admin app.

This commit adds just such a property.

Minor version bump.