Make `has_errors` run faster
The previous implementation of `has_errors` was checking all of the possible
errors before returning. Because checking for errors requires iterating through
all the rows it can be quite expensive.

This commit changes it to return when it finds the first error, which reduces
the number of iterations through the rows in most cases.