Make template change case and space insensitive
When you make changes to a template it doesn’t actually matter if you change
the case or whitespace in a placeholder, eg changing (( FIRSTNAME )) to
((first_name)) is not a breaking change.

This commit changes the `TemplateChange` class to only recognise ‘meaningful’
changes to placeholders, ie changes that modify more than just case or spacing.

It uses the same code for dealing with case/space insensitivity as we do for
looking at the column headers in CSV files to ensure consistency.