Use Jinja to render the email from a file
Having our email template as a string in code was always a hacky,
temporary solution (see 5d0fa6c6a77e2c2ba33b1306b38cb35c8e9d4fd7).

This commit extracts it into a file, which is read on starting the app
(not every time we send an email).

It’s a Jinja2 template, so we can put some logic around it in the
future.

It’s wrapped in a thin class to handle the options which the template
will be able to accept.