Add renderers for plain text email, text messages…
We now have a class to render a HTML email.

This commit adds classes with the same interface to render plain text
emails, text messages, and email and text previews (for use in the
admin app).

By default the `Template` class will return an email or text message
preview when `.replaced` or `.formatted` are called.

Different renderers can be specified by passing them to the `Template`
class when it’s initialised.