Only annotate rows once
Most of our processing of CSV files is based on the `.annotated_rows`
property. `.annotated_rows` takes the raw CSV data and adds some useful
metadata, eg does a row have errors.

This is quite expensive, and we were doing it _loads_ of times.

This commit changes the class to only annotate the rows once, when
the class is initialised.

This cuts processing a 100,000 row file from about 10s to 5s.