Strip HTML from placeholder values
If we are going to preview HTML emails then we need to make sure that
no HTML can be injected into the email beforehand.

This means

1. Stripping HTML from the template (which we do)
2. Stripping HTML from the values of the placeholders (which we weren’t
   doing)

This commit makes sure that placeholders are never replaced with HTML,
in emails (subject and body) or text messages.

---

This is not _yet_ a vulnerability, but will be when we turn on
previewing of templates in the front end.

---

Patch version bump, fixing this bug does not make breaking changes.