Make Markdown take notice of single newlines
Ordinarily Markdown treats single newlines as wrapping, not soft
returns.

So something like this:
> Thanks,
> Government

Comes out as:
> Thanks, Government

We want Markdown to treat newlines as newlines. The proper syntax for
doing this is to add two spaces `  ` at the end of the line, which is
super obscure and unintuitive. So rather than try to document this weird
syntax, let’s just do the conversion for people.