Add caching to the format recipient method
We call format recipient with the same recipient a lot. This is because
we use it to normalize recipients, so for each message in the list we’re
also formatting every message in the whitelist.

Putting a cache in front of this makes it faster, because the whitelist
has a small number of recipients that don’t change.