Make a letter markdown formatter
Turns a letter template into a HTML preview. In the future this might
want to be an image or PDF preview instead, but we’re going with HTML
for quickness.

Letter Markdown formatting is the base type of Markdown formatting, and
works the same way as email used to, except it does nothing with
blockquotes (ie the grey bar on the left).

Email Markdown formatting now inherits from letter markdown formatting
and adds blockquotes.

This is because we won’t be able to do anything like the grey bar with
DVLA.