Make `RecipientCSV` understand letters
Because letters really are a thing[1] now.

This does just enough work for nothing to fall over.

It assumes that an address only has one line (specified by a CSV column
headed ‘address line 1’).

In the future we’ll have to work out how to do multi-field addresses,
pulling from the first `n` columns of the CSV file.

1. https://github.com/alphagov/notifications-admin/pull/201