Make template understand letters
Sets the default renderer for any letter template to be an instance of
`LetterPreview`. This copies the pattern for SMS and email, which is to
default to using the preview renderer.

This is because the preview renderers are used in a bunch of different
templates in the admin app, whereas the for-real renderers tend to only
get used once, at the time the API passes the composed message off to a
provider.