Allow multiple recipient columns
This commit changes the definition of recipient from being one column
to being a list of _n_ columns.

To start with this means a list containing either ‘phone number’ or
‘email address’. This is an internal change that doesn’t result in any
new features.

It is, however, a breaking change because `recipient_column_heading` has
become `recipient_column_headings`, and its type has changed from string
to list.