Define recipient columns for letters
A letter’s recipient, ie the address, has multiple lines, which we are
going to represent as multiple columns in a CSV file.

This commit defines the names of the columns.

N.B. no `name` column, not all address will have a name and it might get
confused with a ((name)) placeholder in the body of the letter.