Validate all recipient columns
If the recipient is in a column called ‘phone number’ or ‘email address’
we do appropriate validation of that column.

If the recipient is spread across multiple columns (like it is for
postal addresses) then we should validate each of those columns
appropriately. For now the appropriate validation is just to check that
the column isn’t empty.