Make values of some address columns optional
An address needs to have at least:
- a first line
- a postcode

If you don’t provide values for any of the other lines then we shouldn’t
consider that an error.

For now the column itself still needs to be present, even if it has a
blank values for every for. This might be something we revisit if it’s
confusing.