Major version bump to 11.0.0
Removes methods from the `Template` and `Recipient` classes (which
probably aren’t being used but would be a breaking change for anything
that is).

Renames methods on the `Template` class that are being used, so
consumers will need updating appropriately. Also removes ones which
are redundant.

Changes the interface of instances of `Renderer` classes.

Changes the `LetterRenderer` to no longer be instantiated with a subject
(it needs to take this from the template instead, so that any
placeholders can be replaced/highlighted).

Changes the preview `Renderer` classes to generate the whole preview.