Pass a dict representation of template to renderer
This lets the renderers cherry-pick whatever attributes of the template
they need in order to contruct the complete message, eg `content` and
`subject`.

This works, and is cleaner, but I think it could be refactored into
something more generic in the future – potentially we’ll have letter
templates that are more complex than just `content` and `subject`.