Remove some methods that are no longer needed
This commit removes some methods from the `Template` class to make it
easier to change, as well as some other tidying up.

The main change is removing `.replaced`/`.formatted` method, and
replacing them with one called `.rendered`.

This will show either the formatted or replaced template content,
depending on whether the template has the values necessary to populate
the placeholders.