Generate complete template previews from renderers
Currently, the `Renderer` classes only generate previews of the body of
the message.

This means that the preview (including showing the recipient, rendering
other wrappers, etc) is split between the admin app and the `utils`
stuff.

This commit brings all that stuff in from the admin app, so that calling
`template.replaced` gives you (for example) and complete preview of an
email, with the from address, service name, recipient’s email, etc. all
populated.