Add a renderer to preview a letter
This new renderer takes a template and returns:

- an `<img>` preview of the letter
- wrapped in a link to the PDF preview of the letter

Generating the URLs manually feels a bit dirty. And probably isn’t going
to be flexible enough when we want to show a preview of a letter with
the placeholders replaced, or a preview of a previous version of a
letter. Works for now as a first pass, so should refactor later.