Move base template tests
The renderers are going to become part of the `Template` class, so
there are going to be more tests of the various types of template than
just the ones in this file.

But I don’t wanna merge them all into one file because they’ll be
testing different things, and it could get very long.