Parametrize markdown tests to expect diff. output
Right now, we expect both the Markdown used in letters and the Markdown
used in emails to convert to the same HTML. But the Markdown in letters
doesn’t need to have all the inline styles, because we’re not working
around the limitations of email clients.

This commit takes the first step towards removing those inline styles,
which is to add extra parametrization to the tests. This way we can test
the same source Markdown, but assert that the output is correct for
letters and emails, even when that output differs.