Remove inline styles from letters HTML
The HTML of our letters is only consumed by the thing that turns it into
a PDF. This means that we don’t have to worry about:
- any issues arising from multiple different clients like we do with
  emails
- any clients which strip out non-inline CSS

So there’s no reason to use inline CSS anymore. This commit removes it
from the parts of the code that convert Markdown into HTML for letters.