Use default markdown renderers where possible
For letters, the Markdown conversion is a lot closer to the default
implementation of Markdown. Where possible we shouldn’t write our own
methods to do the conversion.

So this commit:
- removes these custom conversions where they are functionally
  equivalent to the Markdown default
- does not remove the tests for these conversions, because we still want
  a definition of the output we’re producing (even though this is unit
  testing an external library)

I expected to be able to remove the custom paragraph convertor, but this
isn’t possible because it treats empty lines or strings as paragraphs,
which isn’t behaviour we want.