Add CSS to make a letter look real
This commit adds the CSS to make a letter look like the template that
Sonia and Maria have worked on. The image assets will need to be
deployed to production on admin for this to work (same as we do for
email assets).

Right now it sets the logo to be the HM Government one, which might be
the default for departments that don’t want/need their own branding on
a letter. Hat tip to @harrytrimble for that idea.