Add styling for placeholders in letter templates
This commit highlights placeholders in letter previews the same way
that we do for emails and text messages. The implementation is slightly
different because WeasyPrint doesn’t support box-shadow (see
https://github.com/Kozea/WeasyPrint/pull/149#issuecomment-236052554 for
reasons).

Luckily WeasyPrint _does_ support CSS3 multiple backgrounds, so we can
mask the curved edges with SVG background images instead. This wouldn’t
be a great solution on the web because it you might get a flash of
unstyled placeholder as the images loaded. But because the PDF is
rendered completely before it’s served it shouldn’t be a problem.