Fix email address validation for CSV files
TL;DR We weren’t catching any bad email addresses in CSV files.

So I’m not sure how this ever worked, tried digging back through
`git blame` but even back as far as March 2016 there’s not a change
which is the obvious culprit.

Anyway, the reason is that `BaseException.message` was deprecated in
Python 2.6. Therefore setting `self.message` in the constructor doesn’t
do anything.

The code that catches these exceptions just passes through the error
message. The code that counts which rows have errors treats an `error`
message of `''` on a column equivalant to `None`, ie not an error.

The fix for this is to make sure that the code for setting the default
error message on email addresses passes the message through to the
`Exception` that it’s subclassing.

I’ve also added a more generic test for invalid addresses at the bottom
of a file. This is initially what I thought the problem was, but I think
it’s worth keeping this test because:

- it does catch this specific bug
- it catches a bug which could, potentially, be introduced