Use Nimbus Sans L font to make letter previews
Currently we specify Helvetica Neue as the font to use to generate
letters. This is because when DVLA print the letters, they will be using
Helvetica Neue.

The problem is that Helvetica Neue costs money. Quite a lot of money
[1], when you want to install it on a server. And even then it’s a bit
murky as to whether the licence would cover our usage.

Nimbus Sans L is a font that is very close to Helvetica, and
GPL-licenced [2]. It claims to be metric equivalent, ie the same letters
take up the same amount of space in both typefaces. Testing locally, it
isn’t quite there, but is very close. This means that we can’t guarantee
that our previews will look exactly the same as what DVLA print. But we
could never do that anyway because they’re using different software,
which might composite text differently.

So this commit:

- changes the font used when generating letters
- tweaks the spacing of Nimbus a tiny bit so it looks nicer – this is
  needed because Nimbus doesn’t look to have very good/any kerning pairs

Getting this working:

- locally: install the Nimbus Sans L font on your machine
- on AWS: Andras is working out how to do this

1. About £1800 per year based on
   http://www.myfonts.com/fonts/linotype/neue-helvetica/

 2. https://fontlibrary.org/en/font/nimbus-sans-l