Add a letter output template class
We need to send a notification to DVLA in a format they understand.

This is different to, for example, how we format a letter notification
to preview it.

Therefore this commit adds a new subclass of `Template` whose `__str__`
method will (in subsequent commits) output one line of the file we send
to DVLA.