Render the template conformant to the DVLA spec
The DVLA have specified how they want us to send them each letter to
print. The format is a text file with one row per letter, and various
fields which are pipe delimited.

This commit makes the `LetterDVLATemplate` class render one line of this
file.

The specification for what goes in each column is here:
https://github.com/alphagov/notify-ftp-documentation/blob/bd2164066e3a6b38956db7fb06ba6ced9a2cd50f/dvla_documentation/GDS%20Notify%20-%20DVLA%20File%20Format%20v1_2.xlsx

I’ve copied the contents of this spreadsheet into the tests so that it’s
a bit easier to understand what each field is.

This commit hard codes a lot of the output to match what’s in the
specification. Later commits will make the content dynamic.