Don’t hardcode the dates
The DVLA spec calls for dates in two places. One is for internal use
and one is printed on the letter. They have specified what format the
dates should be in.

The date that we print on the letter should be whatever the date is at
the time of sending. Their example added an extra day to this, but I
think that from the (government) user’s perspective what they see on the
printed letter should match what they see on the preview at the time of
sending.