Don’t hardcode subject line
Like emails, the subject line is a separate piece of data that we give
to the provider. I’m not sure if this is ultimately right, but it works
for now, and means not changing the spec.

Placeholders can be used in the subject line, so I’ve made sure to
change the example to test for this.