Don’t hardcode address block
The address to send the letter to should be whatever the government user
has given to us in their CSV file or API call. The DVLA spec has one
field for each line of the address.

They have allowed seven address lines plus postcode. We only collect
6 address lines, plus postcode. Therefore the 7th line of the address is
hardcoded to be blank. DVLA will strip out any blank lines in the
address before they print the letter.

9a03ae1418ab8f890cc1aaed58620643c2f9ca05 fixed the `Template`s to show
placeholders if their values were missing, rather than an empty string.
This is fine in most cases, but causes a problem with the address block
because some of the placeholder are optional.

This applies more to the `LetterPreviewTemplate`; I’ve done the work for
both here.

The fix is to do a bit more code to work out if someone has supplied
values for the required bit of the address (ie they’re trying to get
a composited preview) or they’re just trying to look at the template
(in which case they should see all the placeholders with the yellow
background). If they’ve only supplied some of the required placeholders
then we should show a combination of the values and the placeholders
with the yellow background.