Strip HTML at Field level
Right now we strip HTML from templates at the point of saving them,
here:
https://github.com/alphagov/notifications-api/blob/d0bea350693685287a161dedaa5df0acb39ee4d3/app/template/rest.py#L44

This also conversts stuff like ampersands to their entity form (eg
`&amp;`) and this is what we save in the database.

This is a bad idea when you’re sending a text message or a letter, in
which an HTML entity makes no sense. But there are still places where we
need to encode or strip HTML to guard against XSS attacks or stop people
putting arbitrary HTML in their email templates.

The right place to do this is in the `Field` class. This is because all
user input to the template is passed through the `Field` class.