Have each `Template` subclass handle HTML
The different types of template need to handle HTML in different ways.
For example an `SMSMessage` should just pass it through so the
characters in the text message match those in the template.

But an `SMSPreview` should encode the entities so that it appears the
same as it does in the textbox where you type the template (and to
mitigate against XSS attacks).

Same reasoning applies to the other template types.