Make subject line render as <h1> not <h2>
In the content of letters, we let people add a heading with the normal
markdown `#` syntax. We convert this to a `<h2>` in the HTML, because
the `<h1>` is reserved for the subject.

Unfortunately, the subject was being inserted by prepending it to the
content of the letter as markdown, and then letting the Markdown
renderer convert it to HTML. But, as described above, the Markdown
renderer was converting it to a `<h2>`. This meant that the type size
for the subject line was too small.

Fixing this means hard-coding the subject to be a `<h1>` in the
template, which is fine because all letters will have a subject, and the
user should have no control over its type size.