Show placeholders on yellow bg if values missing
When all this logic was in the `Template` class we would show all the
placeholders highlighted with a yellow background.

When we moved this logic into the `Field` class, this got broken. If
you provide values for some of the placeholders but not others the ones
without values get replaced with an empty string.

This commit changes the behaviour to show the name of the placeholder,
with the yellow background for _just the placeholder(s) with the missing
value_. I think this is preferable behaviour anyway, because it’s
highlighting which things are still missing.

Minor version bump because this is a bug – although the fix is not quite
the same behaviour as before the interface to this class hasn’t changed.