Refactor numeric_id handling into getter/setter
We’re doing a bunch of checks in here, rather than clutter up the
`__init__` method this code is better separated into its own method, and
a getter/setter is a convenient way of doing this.

It also lets us handle generating the NOTIFICATION_ID from the
`numeric_id`, rather than cluttering up the body of the `__str__` method
with this work.