Fix XSS vulnarability with service name in emails
Service name is user-inputted. When we put it in a text message or email
we should escape it to make sure any HTML that goes into it doesn’t come
render on the page in the admin app.

We already do this for text messages here:
https://github.com/alphagov/notifications-utils/blob/fbb50869caa3b588f9909c8396ff3dc4bf6c04db/notifications_utils/template.py#L162

This commit does the same for emails.

Patch version bump, fixes a bug in a non-breaking way.