Enable markdown style links in emails
We want to do research into how users respond to links which are
‘hidden’ behind some text. The easiest way for us to do this is to send
the emails from Notify. And the easiest way to make this links is using
the default Markdown syntax.

So this commit turns on this feature of Markdown temporarily.

If you are a user of Notify you should not rely on this. It is not a
supported feature. It could stop working at any time.

---

Implementing this also means replacing our hand-crafted auto-linker with
one that uses Markdown to detect links. Two reasons for doing this:

- avoiding awkward conflicts between our linker and Markdown’s link
  syntax

- it’s generally better, for example it handles (http://links.in/brackets)
  properly, which ours doesn’t

The only reason why we didn’t use Markdown’s autolinker in the first
place is because were weren’t using Markdown at all when we added
autolinking to email templates.

Link detection was added on 6 June 2016:
https://github.com/alphagov/notifications-utils/commit/a66383a0839b7545ae8c3b7e354a59c65caf3920

Markdown was added 15 July 2016:
https://github.com/alphagov/notifications-utils/pull/54/commits/df9f167e6bc201cbffcf1d7447800504283f8ffa