convert non-gsm characters to gsm-compatible versions in text messages
GSM is an ascii-like (but not ascii) character set that text messages
are typically sent in. The providers we work with handle non-GSM chars
differently - one downgrades non-gsm to GSM compatible, and one sends
in unicode encoding, which takes three times as much space, and so
changes the amount billable.

This commit changes SMSTemplate to downgrade the content to GSM before
printing. For implementation details, read the docstrings in gsm.py