Pass contact block through to preview
This commit takes any string passed to `LetterPreviewTemplate` as a
contact block and generates a preview with this information in the
top-right of the rendered letter (according to the dimensions on Sonia
and Maria’s design).

Because the preview is generated from HTML, we have to convert newlines
to HTML `<br>`s.