Check for presence of recipient before validating
If you have missed out a phone number in your CSV file, it doesn’t make
sense for us to tell you that <empty cell> is ‘Not a UK mobile number’.
It makes more sense for us to tell you that the data is missing from
that cell (which is what we do when you have empty cells in other
columns).

This came up because a user uploaded a file which had a blank row
between the column headers and the data, and couldn’t understand why
their file wasn’t being accepted by Notify.